requirements.txt
aiofiles==0.7.0
aiosqlite==0.17.0
anyio==3.2.1
asgiref==3.3.4
authentication==1.1.0
bokeh==2.3.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
dashboard==0.0.6
databases==0.4.3
Django==3.2.4
django-cors-headers==3.7.0
django-csp=3.7
django-email-users==1.1.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
geographiclib==1.52
geopy==2.1.0
greenlet==1.1.0
h11==0.12.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mysqlclient==2.0.1
numpy==1.21.0
orm==0.2.0.dev1
packaging==20.9
pandas==1.2.5
passlib==1.7.4
Pillow==8.2.0
pycryptodome==3.10.1
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
six==1.16.0
sniffio==1.2.0
South==1.0.2
SQLAlchemy==1.3.24
sqlparse==0.4.1
starlette==0.15.0
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
typesystem==0.3.0.dev0
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.14.0
Werkzeug==2.0.1
wincertstore==0.2
WTForms==2.3.3
